+++ services/tracker/src/tracker/benchmark_service.py
@@ -1,39 +1,38 @@
+import logging
 import os
 from typing import Any
 
 import httpx
 from daytona import AsyncDaytona, DaytonaConfig
-from dotenv import load_dotenv
 
 from tracker.exceptions import BenchmarkServiceError
-from tracker.logger import get_logger
 
-load_dotenv()
-
-logger = get_logger(__name__)
+logger = logging.getLogger(__name__)
 
 
 class BenchmarkService:
     _name: str
     _url: str
     _environment_keys: dict[str, str]
+    _daytona_client: AsyncDaytona
 
     def __init__(self, name: str, url: str):
-        logger.info(f"Initializing benchmark service for {name} at {url}")
-
         self._name = name
         self._url = url
         self._environment_keys = self.daytona_keys()
+        self._daytona_client = self.create_daytona_client()
 
-        self._daytona = AsyncDaytona(
-            config=DaytonaConfig(
-                api_key=self._environment_keys["DAYTONA_API_KEY"],
-                api_url=self._environment_keys["DAYTONA_API_URL"],
-                target=self._environment_keys["DAYTONA_TARGET"],
-            )
-        )
+    @property
+    def name(self) -> str:
+        return self._name
 
-        logger.info(f"Benchmark service initialized for {name} at {url}")
+    @property
+    def environment_keys(self) -> dict[str, str]:
+        return self._environment_keys
+
+    @property
+    def daytona_client(self) -> AsyncDaytona:
+        return self._daytona_client
 
     @staticmethod
     def daytona_keys() -> dict[str, str]:
@@ -43,21 +42,31 @@ class BenchmarkService:
             "DAYTONA_TARGET": os.getenv("DAYTONA_TARGET") or "",
         }
 
-        missing_keys = [key for key, value in environment_keys.items() if not value]
+        missing_keys: list[str] = []
+        for key, value in environment_keys.items():
+            if not value:
+                missing_keys.append(key)
 
         if missing_keys:
-            raise BenchmarkServiceError(
+            raise ValueError(
                 f"The following environment variables are not set: {', '.join(missing_keys)}. Please set them in your `.env` file so that they can be sourced."
             )
 
         return environment_keys
 
+    def create_daytona_client(self) -> AsyncDaytona:
+        return AsyncDaytona(
+            config=DaytonaConfig(
+                api_key=self.environment_keys["DAYTONA_API_KEY"],
+                api_url=self.environment_keys["DAYTONA_API_URL"],
+                target=self.environment_keys["DAYTONA_TARGET"],
+            )
+        )
+
     async def request_health_check(self) -> dict[str, str]:
         """
         Requests health check from benchmark service
         """
-        logger.info(f"Performing health check for {self._name} benchmark service")
-
         async with httpx.AsyncClient() as client:
             response = await client.get(f"{self._url}/health")
 
@@ -68,32 +77,28 @@ class BenchmarkService:
                 f"Health check failed with status code {response.status_code}, response: {response.text}"
             )
 
-        logger.info(f"Health check passed for {self._name} benchmark service")
-
         return response.json()
 
-    async def request_verify_task_ids(self, task_ids: list[str]) -> list[str]:
+    async def request_verify_task_ids(self, task_ids: list[str] | None) -> list[str]:
         """
         Requests verify task ids from benchmark service
-        Returns a list of verified task ids
         """
-        logger.info(f"Verifying task IDs: {task_ids}")
 
-        params = {"task_ids": task_ids}
+        params: dict[str, list[str]] = {}
+        if task_ids is not None:
+            params["task_ids"] = task_ids
+
         async with httpx.AsyncClient() as client:
             response = await client.get(f"{self._url}/verify-task-ids", params=params)
+
+        logger.debug(f"Verify task ids response: {response.json()}")
 
         if response.status_code != 200:
             raise BenchmarkServiceError(
                 f"Verify task ids failed with status code {response.status_code}, response: {response.text}"
             )
 
-        response_json = response.json()
-        verified_task_ids: list[str] = response_json["task_ids"]
-
-        logger.info(f"Task IDs verified successfully: {verified_task_ids}")
-
-        return verified_task_ids
+        return response.json().get("task_ids", [])
 
     async def request_retrieve_tasks(
         self, task_ids: list[str], skip_validation: bool = False
@@ -101,25 +106,24 @@ class BenchmarkService:
         """
         Requests retrieve tasks from benchmark service
         """
-        logger.info(f"Retrieving tasks for verified task IDs: {task_ids}")
 
         params = {"task_ids": task_ids, "skip_validation": skip_validation}
         async with httpx.AsyncClient() as client:
             response = await client.get(f"{self._url}/retrieve-tasks", params=params)
+
+        logger.debug(f"Retrieve tasks response: {response.json()}")
 
         if response.status_code != 200:
             raise BenchmarkServiceError(
                 f"Retrieve tasks failed with status code {response.status_code}, response: {response.text}"
             )
 
-        logger.info("Tasks retrieved successfully")
         return response.json()
 
     async def request_setup_task(self, task_id: str, instance_id: str) -> dict[str, str]:
         """
         Requests setup task from benchmark service
         """
-        logger.info(f"Setting up task {task_id} with instance {instance_id}")
 
         async with httpx.AsyncClient() as client:
             response = await client.post(
@@ -133,12 +137,12 @@ class BenchmarkService:
                 },
             )
 
+        logger.debug(f"Setup task response: {response.json()}")
+
         if response.status_code != 200:
             raise BenchmarkServiceError(
                 f"Setup task failed with status code {response.status_code}, response: {response.text}"
             )
-
-        logger.info(f"Task {task_id} setup successfully")
 
         return response.json()
 
@@ -146,7 +150,6 @@ class BenchmarkService:
         """
         Requests evaluate instance from benchmark service
         """
-        logger.info(f"Evaluating instance {instance_id} for task {task_id}")
 
         async with httpx.AsyncClient() as client:
             response = await client.post(
@@ -160,21 +163,19 @@ class BenchmarkService:
                 },
             )
 
+        logger.debug(f"Evaluate instance response: {response.json()}")
+
         if response.status_code != 200:
             raise BenchmarkServiceError(
                 f"Evaluate instance failed with status code {response.status_code}, response: {response.text}"
             )
 
-        logger.info(f"Instance {instance_id} evaluated successfully")
-
         return response.json()
 
-    async def request_final_score(self, evaluation_results: dict[str, dict[str, Any]]) -> dict[str, str]:
+    async def request_final_score(self, evaluation_results: dict[str, dict[str, Any]]) -> dict[str, Any]:
         """
         Requests final score from benchmark service
         """
-        logger.info(f"Producing final score for tasks {evaluation_results.keys()}")
-
         async with httpx.AsyncClient() as client:
             response = await client.post(
                 f"{self._url}/final-score",
@@ -182,18 +183,11 @@ class BenchmarkService:
                 headers={"Content-Type": "application/json"},
             )
 
+        logger.debug(f"Final score response: {response.json()}")
+
         if response.status_code != 200:
             raise BenchmarkServiceError(
                 f"Final score failed with status code {response.status_code}, response: {response.text}"
             )
 
-        logger.info("Final score produced successfully")
-
         return response.json()
-
-    @property
-    def daytona(self) -> AsyncDaytona:
-        """
-        Returns the Daytona client instance.
-        """
-        return self._daytona
